# probe-service/src/main/java/com/tum/in/cm/probeservice/web/ws/dto/HttpMeasurementSpecification.java
package com.tum.in.cm.probeservice.web.ws.dto;

import com.tum.in.cm.probeservice.util.Constants;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
public class HttpMeasurementSpecification extends MeasurementSpecification {
    private String target;
    private Constants.HttpMethod method;
    private String queryString;
}